backend/requirements.txt
 
django
djangorestframework
python-dotenv
requests
google-generativeai
PyPDF2
pydantic
drf_spectacular